Install Java

+++ library/introduce.sh
@@ -67,44 +67,51 @@ fi
 
 # Outils de travail
 if [ ${wsd_project_environment} == "development" ]; then
-    # Todo : si en DEV : proposer d'installer : java, phpstorm, smartgit...
-    echo 'java, phpstorm, smartgit...';
+    echo "Installer Java, phpStorm, SmartGit... ?" && read REPLY && if [ "$REPLY" == "y" ]; then
+        sudo echo "deb http://ppa.launchpad.net/webupd8team/java/ubuntu trusty main" | tee /etc/apt/sources.list.d/webupd8team-java.list;
+        sudo echo "deb-src http://ppa.launchpad.net/webupd8team/java/ubuntu trusty main" | tee -a /etc/apt/sources.list.d/webupd8team-java.list;
+        sudo apt-key adv --keyserver hkp://keyserver.ubuntu.com:80 --recv-keys EEA14886;
+        sudo apt-get update -y;
+        sudo apt-get -y install oracle-java8-installer;
+        # java, phpstorm, smartgit...
+    fi
 fi
 
 # Installation de NeoVIM
-echo "Installer NeoVIM ?" && read REPLY && if [ "$REPLY" == "y" ]; then
-    cd ~;
-    sudo apt-get install libtool autoconf automake cmake libncurses5-dev g++ pkg-config;
-    wget https://github.com/neovim/neovim/archive/master.zip;
-    unzip master.zip;
-    cd neovim-master;
-    make;
-    cmake -DCMAKE_INSTALL_PREFIX:PATH=/usr/;
-    sudo make install;
-    cd ~;
-    rm -rf neovim-master;
-    rm -f master.zip;
-    mkdir -p ~/.nvim/bundle ~/.nvim/autoload;
-    touch ~/.nvimrc;
+if [ ${wsd_project_environment} != "development" ]; then
+    echo "Installer NeoVIM ?" && read REPLY && if [ "$REPLY" == "y" ]; then
+        cd ~;
+        sudo apt-get install libtool autoconf automake cmake libncurses5-dev g++ pkg-config;
+        wget https://github.com/neovim/neovim/archive/master.zip;
+        unzip master.zip;
+        cd neovim-master;
+        make;
+        cmake -DCMAKE_INSTALL_PREFIX:PATH=/usr/;
+        sudo make install;
+        cd ~;
+        rm -rf neovim-master;
+        rm -f master.zip;
+        mkdir -p ~/.nvim/bundle ~/.nvim/autoload;
+        touch ~/.nvimrc;
 
-    # Plugin Pathogen
-    curl -LSso ~/.nvim/autoload/pathogen.vim https://tpo.pe/pathogen.vim;
-    echo 'execute pathogen#infect()' >> ~/.nvimrc;
-    echo 'syntax on' >> ~/.nvimrc;
-    echo 'filetype plugin indent on' >> ~/.nvimrc;
+        # Plugin Pathogen
+        curl -LSso ~/.nvim/autoload/pathogen.vim https://tpo.pe/pathogen.vim;
+        echo 'execute pathogen#infect()' >> ~/.nvimrc;
+        echo 'syntax on' >> ~/.nvimrc;
+        echo 'filetype plugin indent on' >> ~/.nvimrc;
 
-    # Plugin NERD Tree
-    cd ~/.nvim/bundle && git clone https://github.com/scrooloose/nerdtree.git;
-    echo 'autocmd VimEnter * NERDTree' >> ~/.nvimrc;
+        # Plugin NERD Tree
+        cd ~/.nvim/bundle && git clone https://github.com/scrooloose/nerdtree.git;
+        echo 'autocmd VimEnter * NERDTree' >> ~/.nvimrc;
 
-    # Plugin VIM Markdown
-    cd ~/.nvim/bundle && git clone https://github.com/plasticboy/vim-markdown.git;
+        # Plugin VIM Markdown
+        cd ~/.nvim/bundle && git clone https://github.com/plasticboy/vim-markdown.git;
 
-    # ColorScheme
-    cd ~/.nvim/bundle && git clone https://github.com/morhetz/gruvbox.git;
-    echo 'colorscheme gruvbox' >> ~/.nvimrc;
+        # ColorScheme
+        cd ~/.nvim/bundle && git clone https://github.com/morhetz/gruvbox.git;
+        echo 'colorscheme gruvbox' >> ~/.nvimrc;
 
-    # Désinstallation de nano
-    sudo apt-get -y remove nano vim;
-   
+        # Désinstallation de nano
+        sudo apt-get -y remove nano vim;
+    fi
 fi
